refactor: testutils package
