+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/collection/CollectionExtTest.kt
@@ -1,6 +1,5 @@
 package xyz.lbres.kotlinutils.collection
 
-import xyz.lbres.kotlinutils.assertEmpty
 import xyz.lbres.kotlinutils.collection.list.IntList
 import xyz.lbres.kotlinutils.collection.list.StringList
 import xyz.lbres.kotlinutils.collection.multiset.ConstMultiSet
@@ -15,6 +14,7 @@ import xyz.lbres.kotlinutils.collection.multiset.multiSetOf
 import xyz.lbres.kotlinutils.collection.multiset.mutableMultiSetOf
 import xyz.lbres.kotlinutils.internal.constants.Suppressions
 import xyz.lbres.kotlinutils.number.isZero
+import xyz.lbres.kotlinutils.testutils.assertEmpty
 import kotlin.math.max
 import kotlin.math.min
 import kotlin.test.Test
