+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/collection/MutableCollectionExtTest.kt
@@ -1,9 +1,9 @@
 package xyz.lbres.kotlinutils.collection
 
-import xyz.lbres.kotlinutils.CompList
 import xyz.lbres.kotlinutils.collection.multiset.MutableMultiSet
 import xyz.lbres.kotlinutils.collection.multiset.mutableMultiSetOf
-import xyz.lbres.kotlinutils.runTestWithRetry
+import xyz.lbres.kotlinutils.testutils.CompList
+import xyz.lbres.kotlinutils.testutils.runTestWithRetry
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNotEquals
