+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/collection/multiset/MultiSetCompanionTest.kt
@@ -1,9 +1,9 @@
 package xyz.lbres.kotlinutils.collection.multiset
 
-import xyz.lbres.kotlinutils.CompList
 import xyz.lbres.kotlinutils.collection.list.IntList
 import xyz.lbres.kotlinutils.collection.multiset.impl.ConstMultiSetImpl
 import xyz.lbres.kotlinutils.collection.multiset.impl.ConstMutableMultiSetImpl
+import xyz.lbres.kotlinutils.testutils.CompList
 import kotlin.test.Test
 import kotlin.test.assertEquals
