+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/collection/multiset/testutils/commonMinusTests.kt
@@ -1,12 +1,12 @@
 package xyz.lbres.kotlinutils.collection.multiset.testutils
 
-import xyz.lbres.kotlinutils.CompList
-import xyz.lbres.kotlinutils.assertEmpty
 import xyz.lbres.kotlinutils.collection.list.IntList
 import xyz.lbres.kotlinutils.collection.multiset.ConstMultiSet
 import xyz.lbres.kotlinutils.collection.multiset.MultiSet
 import xyz.lbres.kotlinutils.collection.multiset.multiSetOf
 import xyz.lbres.kotlinutils.internal.constants.Suppressions
+import xyz.lbres.kotlinutils.testutils.CompList
+import xyz.lbres.kotlinutils.testutils.assertEmpty
 import xyz.lbres.kotlinutils.utils.simpleIf
 import kotlin.test.assertEquals
